fix(app): split profile route into layout and index to support public profiles
Refactor profile routing so /profile acts as a layout route rendering an Outlet, moving the authenticated dashboard to /profile.index.tsx. This allows /profile/$username to render the public profile properly. Also adds an Edit Profile action for the profile owner when viewing their own public profile.
